refactor: MessageScreen file 위치 변경
